Refactors survive third turn event to second turn
Corrects the timing of the "survive with evade" and
"survive with guard" events to trigger on the second turn
instead of the third turn.
