+++ src/js/custom-battle-events/survive-super-power-with-guard/procedures/before-last-state/should-play-survive-sedond-turn-with-guard.ts
