+++ src/js/custom-battle-events/survive-super-power-with-guard/stories/survive-second-turn-with-guard.ts
